+++ cv_personal_page/frontend/javascript/navigation_bar.js
@@ -5,9 +5,8 @@ export const navigationBar = document.querySelector("#nav-bar");
 // Control of navigation menu buttons
 const navigationMenuButton = document.querySelector("#navigation-menu-button");
 const navigationMenu = document.querySelector("#navigation-menu");
-export const navigationMenuLinks = document.querySelectorAll(
-  "#navigation-menu > ul > li > a",
-);
+export const navigationMenuLinks =
+  document.querySelectorAll("#navigation-menu a");
 
 const maxMdHiddenClassName = "max-md:hidden";